Fixed an event subscription memory leak
QueryLedgerEntries was called too many times the more times the page was opened. Turns out, finalizers aren't called when VM is disposed. In the comments of this SO thread (https://stackoverflow.com/questions/76209023/when-to-unsubscribe-from-events-in-content-views) someone pointed out that the reason why finalizers aren't called might be because VM won't be garbage collected because the EventHandler is still referencing their method delegate.
The Shell.BackButtonBehaviour XAML tag comes from this Microsoft Learn question: https://learn.microsoft.com/en-us/answers/questions/1002088/why-the-pages-destructor-never-called.

Amended. It wasn't fully fixed, it is now. But (oh the horror!) I had to add code directly to the modal class of the View. I sure love me some MAUI jank... https://learn.microsoft.com/en-us/dotnet/api/xamarin.forms.page.onbackbuttonpressed?view=xamarin-forms - This is the method I had to override, because the XAML tag works... only on the Return button visible on the navigation tab at the top of the page, not the Android's built-in Return touchbutton...