Revert "Changed LedgerPage to make use of EmptyView instead of seperate bool property"
Reverting this commit, because CollectionView is bugged and doesn't show its EmptyView upon loading the LedgerPage. It shows properly once LedgerEntries are filled and cleared. Workaround of adding an empty BalanceLedger object, then removing it did not work. According to a question, linked below, this bug was already present in Xamarin.Android.

https://stackoverflow.com/a/70711123/21342746

This reverts commit 79d48b65255f7456c0bc2e2d962e973416bf4ab3.